refactor: split app logic into separate file
Move the majority of the application's logic from `main.rs` into
`lib.rs` to faciliate the first move towards encapsulation and testing.
